% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_functions.R
\name{catalog_tables}
\alias{catalog_tables}
\title{Return available catalog tables}
\usage{
catalog_tables(con = dbcon)
}
\arguments{
\item{con}{sqlite connection, default is dbcon}
}
\description{
Function will return a list of tables that are available
in the sqlite connection
}
