% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_functions.R
\name{column_info}
\alias{column_info}
\title{return column information}
\usage{
column_info(column, con = dbcon)
}
\arguments{
\item{column}{string name of column to find}

\item{con}{connection to the sqllite db, default is dbcon}
}
\description{
Function will return information from master table
in the sqlite db
}
