% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_functions.R
\name{get_catalog_connection}
\alias{get_catalog_connection}
\title{Function to get sql list version of pmap data catalog}
\usage{
get_catalog_connection(dbpath)
}
\arguments{
\item{dbpath}{string path to the sql list db}
}
\description{
Will return a connection to the pmap data catalog
}
