% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_functions.R
\name{table_info}
\alias{table_info}
\title{return table information}
\usage{
table_info(table, con = dbcon)
}
\arguments{
\item{table}{string name of table to find}

\item{con}{connection to the sqllite db, default is dbcon}
}
\description{
Function will return information from master table
in the sqlite db given a table  name
}
